Version v3.0
This is the third version of my chess engine. It has recieved some
significant changes from v2 and is now considerably smarter. It is
approximately 1400-1600 elo and will hopefully continue to become
stronger. I have plans on optimizing the engine even more in the future
but these changes have made quite the improvement already.