Version v4.0
This version is a large improvement over version 3. It implements
pondering, multithreading, improved tt for multiple threadsd, and many more updates. However, there are
still some major improvements required. The search is inefficient and
needs to be refactored, so there is still a lot of work to go but this
is a large improvement compared to the previous versions.